feat(redesign-list-pages): create separate component for page limit dropdown
[#159245473]
