+++ src/_utils/pageRowOptions.js
@@ -1,0 +1,16 @@
+const rowOptions = [
+  {
+    text: '10 Rows',
+    value: 10
+  },
+  {
+    text: '20 Rows',
+    value: 20
+  },
+  {
+    text: '30 Rows',
+    value: 30
+  }
+];
+
+export default rowOptions;
